Create bulletproof glass blocks
Since bullets can now break glass, I've created bulletproof glass, along with stained variants. There are no bulletproof panes, not yet at least.

Update mods.toml with proper license information, and update the GitHub URLs to reflect the repository name change

+++ src/main/java/com/anonymoushacker1279/immersiveweapons/entity/projectile/BulletEntity.java
@@ -346,7 +346,7 @@ public class BulletEntity {
 			this.setShotFromCrossbow(false);
 			this.func_213870_w();
 			
-			if (canBreakGlass && !this.hasAlreadyBrokeGlass && this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass")) || this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass_panes"))) {;
+			if (canBreakGlass && !this.hasAlreadyBrokeGlass && !this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:bulletproof_glass")) && this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass")) || this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass_panes"))) {;
 				this.world.destroyBlock(blockStateRayTraceResult.getPos(), false);
 				this.hasAlreadyBrokeGlass = true;
 			}
@@ -666,7 +666,7 @@ public class BulletEntity {
 			this.setShotFromCrossbow(false);
 			this.func_213870_w();
 			
-			if (canBreakGlass && !this.hasAlreadyBrokeGlass && this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass")) || this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass_panes"))) {;
+			if (canBreakGlass && !this.hasAlreadyBrokeGlass && !this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:bulletproof_glass")) && this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass")) || this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass_panes"))) {;
 				this.world.destroyBlock(blockStateRayTraceResult.getPos(), false);
 				this.hasAlreadyBrokeGlass = true;
 			}
@@ -986,7 +986,7 @@ public class BulletEntity {
 			this.setShotFromCrossbow(false);
 			this.func_213870_w();
 			
-			if (canBreakGlass && !this.hasAlreadyBrokeGlass && this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass")) || this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass_panes"))) {;
+			if (canBreakGlass && !this.hasAlreadyBrokeGlass && !this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:bulletproof_glass")) && this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass")) || this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass_panes"))) {;
 				this.world.destroyBlock(blockStateRayTraceResult.getPos(), false);
 				this.hasAlreadyBrokeGlass = true;
 			}
@@ -1307,7 +1307,7 @@ public class BulletEntity {
 			this.setShotFromCrossbow(false);
 			this.func_213870_w();
 			
-			if (canBreakGlass && !this.hasAlreadyBrokeGlass && this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass")) || this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass_panes"))) {;
+			if (canBreakGlass && !this.hasAlreadyBrokeGlass && !this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:bulletproof_glass")) && this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass")) || this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass_panes"))) {;
 				this.world.destroyBlock(blockStateRayTraceResult.getPos(), false);
 				this.hasAlreadyBrokeGlass = true;
 			}
@@ -1627,7 +1627,7 @@ public class BulletEntity {
 			this.setShotFromCrossbow(false);
 			this.func_213870_w();
 			
-			if (canBreakGlass && !this.hasAlreadyBrokeGlass && this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass")) || this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass_panes"))) {;
+			if (canBreakGlass && !this.hasAlreadyBrokeGlass && !this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:bulletproof_glass")) && this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass")) || this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass_panes"))) {;
 				this.world.destroyBlock(blockStateRayTraceResult.getPos(), false);
 				this.hasAlreadyBrokeGlass = true;
 			}
@@ -1948,7 +1948,7 @@ public class BulletEntity {
 			this.setShotFromCrossbow(false);
 			this.func_213870_w();
 			
-			if (canBreakGlass && !this.hasAlreadyBrokeGlass && this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass")) || this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass_panes"))) {;
+			if (canBreakGlass && !this.hasAlreadyBrokeGlass && !this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:bulletproof_glass")) && this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass")) || this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass_panes"))) {;
 				this.world.destroyBlock(blockStateRayTraceResult.getPos(), false);
 				this.hasAlreadyBrokeGlass = true;
 			}
@@ -2270,7 +2270,7 @@ public class BulletEntity {
 			this.setShotFromCrossbow(false);
 			this.func_213870_w();
 			
-			if (canBreakGlass && !this.hasAlreadyBrokeGlass && this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass")) || this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass_panes"))) {;
+			if (canBreakGlass && !this.hasAlreadyBrokeGlass && !this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:bulletproof_glass")) && this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass")) || this.inBlockState.getBlock().isIn(BlockTags.makeWrapperTag("forge:glass_panes"))) {
 				this.world.destroyBlock(blockStateRayTraceResult.getPos(), false);
 				this.hasAlreadyBrokeGlass = true;
 			}
